refactor(screen_files):reorder files and folders, deleting unnecesary files
