--- lib/navigation/prev_version/questions.dart
@@ -1,104 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:wisdrive/widgets/answer_botton.dart';
-import 'package:wisdrive/data/app_theme.dart';
-import 'package:wisdrive/data/questions.dart';
-import 'package:google_fonts/google_fonts.dart';
-
-class QuestionsScreen extends StatefulWidget {
-  const QuestionsScreen({super.key, required this.onSelectAnswer});
-
-  final void Function(String answer) onSelectAnswer; //function called in Quiz
-
-  @override
-  State<QuestionsScreen> createState() {
-    return _QuestionsScreenState();
-  }
-}
-
-class _QuestionsScreenState extends State<QuestionsScreen> {
-  var currentQuestionIndex = 0;
-
-  void answerQuestion(String selectedAnswer) {
-    widget.onSelectAnswer(selectedAnswer);
-    setState(() {
-      currentQuestionIndex++;
-    });
-  }
-
-  @override
-  Widget build(context) {
-    final currentQuestion = questions[currentQuestionIndex];
-
-    return SizedBox(
-      //SizedBox is used as a Center Widget
-      width: double.infinity,
-      child: Container(
-        padding: const EdgeInsets.all(30),
-        margin: const EdgeInsets.only(top: 100), //margen superior
-        decoration: const BoxDecoration(
-          color: white,
-          borderRadius: BorderRadius.only(
-            topLeft: Radius.elliptical(50, 30),
-            topRight: Radius.elliptical(50, 30),
-          ),
-        ),
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          crossAxisAlignment: CrossAxisAlignment.stretch,
-          children: [
-            Container(
-              padding: const EdgeInsets.all(30),
-              decoration: BoxDecoration(
-                borderRadius: BorderRadius.circular(20),
-                gradient: const LinearGradient(
-                  colors: [
-                    blackPurple,
-                    royalPurple,
-                  ],
-                  begin: Alignment.topCenter,
-                  end: Alignment.bottomCenter,
-                ),
-              ),
-              child: Text(
-                currentQuestion.text,
-                style: GoogleFonts.poppins(
-                  color: white,
-                  fontSize: 24,
-                  fontWeight: FontWeight.bold,
-                ),
-                textAlign: TextAlign.center,
-              ),
-            ),
-            const SizedBox(height: 30),
-            //Return each shuffled answer inside QuizQuestion array and give each a function
-            ...currentQuestion.getShuffledAnswers().map((answer) {
-              return AnswerButton(
-                answerText: answer,
-                onTap: () {
-                  answerQuestion(answer);
-                },
-              );
-            }),
-            const SizedBox(height: 100),
-            ElevatedButton(
-              onPressed: () {},
-              style: ElevatedButton.styleFrom(
-                backgroundColor: purple,
-                padding: const EdgeInsets.all(10),
-              ),
-              child: Text(
-                'CONTESTAR',
-                textAlign: TextAlign.center,
-                style: GoogleFonts.poppins(
-                  color: white,
-                  fontSize: 24,
-                  fontWeight: FontWeight.bold,
-                ),
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
