+++ lib/navigation/screens/quizes/questions_screen.dart
@@ -5,7 +5,6 @@ import 'package:supabase_flutter/supabase_flutter.dart';
 import 'package:wisdrive/constraints/helper_functions.dart';
 import 'package:wisdrive/controllers/theme_controller.dart';
 import 'package:wisdrive/data/app_theme.dart';
-import 'package:wisdrive/navigation/screens/home/quizes_screen.dart';
 import 'package:wisdrive/widgets/home/sidebar_menu.dart';
 
 class QuestionsScreen extends StatefulWidget {
