+++ lib/widgets/home/module_list.dart
@@ -5,7 +5,7 @@ import 'package:wisdrive/controllers/theme_controller.dart';
 import 'package:get/get.dart';
 import 'package:wisdrive/data/app_theme.dart';
 import 'package:wisdrive/generated/l10n.dart';
-import 'package:wisdrive/navigation/screens/home/quizes_screen.dart';
+import 'package:wisdrive/navigation/screens/quizes/quizes_screen.dart';
 
 class ModuleList extends StatelessWidget {
   final int selectedCategoryId;
